refactor: avoid unhandledRejection in useLoadable
useSyncExternalStore in some cases may cause some Promises to be generated without a chance to be caught by `.catch`, ultimately leading to an unhandledRejection in React applications.

This commit passes by checking whether the previously generated promise in useSyncExternalStore was correctly handled, wrapping those Promises that did not have a chance to be processed with a callback function that silently handles all exceptions.

Co-authored-by: Yuchen Zhang <zhangyc@kanyun.com>